Начал писать клиент, пока только парсинг командной строки
